+++ src/Components/ClustersListTable/ClustersListTable.js
@@ -37,6 +37,8 @@ import {
   mapClustersToRows,
   paramParser,
   passFiltersCluster,
+  removeFilterParam as _removeFilterParam,
+  addFilterParam as _addFilterParam,
   translateSortParams,
   updateSearchParams,
 } from '../Common/Tables';
@@ -69,6 +71,12 @@ const ClustersListTable = ({
   const loadingState = isUninitialized || isFetching || !rowsFiltered;
   const errorState = isError;
   const successState = isSuccess;
+
+  const removeFilterParam = (param) =>
+    _removeFilterParam(filters, updateFilters, param);
+
+  const addFilterParam = (param, values) =>
+    _addFilterParam(filters, updateFilters, param, values);
 
   useEffect(() => {
     setDisplayedRows(
@@ -150,25 +158,6 @@ const ClustersListTable = ({
     );
   };
 
-  const removeFilterParam = (param) => {
-    const { [param]: omitted, ...newFilters } = { ...filters, offset: 0 };
-    updateFilters({
-      ...newFilters,
-      ...(param === 'text'
-        ? { text: '' }
-        : param === 'hits'
-        ? { hits: [] }
-        : {}),
-    });
-  };
-
-  // TODO: update URL when filters changed
-  const addFilterParam = (param, values) => {
-    values.length > 0
-      ? updateFilters({ ...filters, offset: 0, ...{ [param]: values } })
-      : removeFilterParam(param);
-  };
-
   const filterConfigItems = [
     {
       label: intl.formatMessage(messages.name).toLowerCase(),
@@ -176,7 +165,7 @@ const ClustersListTable = ({
         key: 'text-filter',
         onChange: (_event, value) => updateFilters({ ...filters, text: value }),
         value: filters.text,
-        placeholder: intl.formatMessage(messages.filterBy),
+        placeholder: intl.formatMessage(messages.filterByName),
       },
     },
     {